Driver Manager refactoring

+++ OS/src/hal/generic/mmu/mmu.h
@@ -13,7 +13,7 @@
 #include "../../../kernel/process/process.h"
 
 mmu_table_t* mmu_init_process(uint8_t kernel);
-unsigned int mmu_map(mmu_table_t* table, void* virtual_address, void* physical_address);
+size_t mmu_map(mmu_table_t* table, void* virtual_address, void* physical_address);
 
 void mmu_init_hal(void);
 void mmu_activate_process(Process_t* process);
